Move terminal velocities test to the same file
